refactor: v1.6.4 — rename rates.py → shared.py, deduplicate helpers
Consolidate _num, _sanitize, f_dur, f_tok, f_cost into shared.py
(previously duplicated across statusline.py and monitor.py).
Update all imports, docs, CI workflows, and templates.
