+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -5,7 +5,7 @@
 ## Checklist
 
 - [ ] `python3 tests.py` (or `py tests.py` on Windows) passes (all tests green)
-- [ ] All files compile: `python3 -c "import py_compile; [py_compile.compile(f, doraise=True) for f in ('rates.py','statusline.py','monitor.py','update.py')]"` (use `py` on Windows)
+- [ ] All files compile: `python3 -c "import py_compile; [py_compile.compile(f, doraise=True) for f in ('shared.py','statusline.py','monitor.py','update.py')]"` (use `py` on Windows)
 - [ ] Tested manually with a live Claude Code session
 - [ ] Zero new dependencies introduced
 - [ ] `MAX_FILE_SIZE` and ANSI palette kept in sync across `statusline.py` and `monitor.py` (if changed)
